requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
langchain-chroma>=0.0.5
chromadb<0.5.0
python-dotenv>=1.0.0
pandas>=2.2.0
openai>=1.12.0
setuptools>=65.5.1
wheel>=0.38.4
build>=0.7.0
hnswlib>=0.7.0
typing-inspect>=0.8.0
typing_extensions>=4.5.0
SQLAlchemy>=2.0.0
fastapi>=0.95.2
uvicorn>=0.15.0 